// contracts/typescript/models/ProvisionMemberRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ProvisionMemberRequest
 */
export interface ProvisionMemberRequest {
    /**
     *
     */
    displayName: string;
    /**
     *
     */
    email: string;
    /**
     *
     */
    isAccountAdministrator: boolean;
    /**
     *
     */
    isEmployee: boolean;
    /**
     *
     */
    isManager: boolean;
}

/**
 * Check if a given object implements the ProvisionMemberRequest interface.
 */
export function instanceOfProvisionMemberRequest(value: object): value is ProvisionMemberRequest {
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('isAccountAdministrator' in value) || value['isAccountAdministrator'] === undefined) return false;
    if (!('isEmployee' in value) || value['isEmployee'] === undefined) return false;
    if (!('isManager' in value) || value['isManager'] === undefined) return false;
    return true;
}

export function ProvisionMemberRequestFromJSON(json: any): ProvisionMemberRequest {
    return ProvisionMemberRequestFromJSONTyped(json, false);
}

export function ProvisionMemberRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProvisionMemberRequest {
    if (json == null) {
        return json;
    }
    return {

        'displayName': json['displayName'],
        'email': json['email'],
        'isAccountAdministrator': json['isAccountAdministrator'],
        'isEmployee': json['isEmployee'],
        'isManager': json['isManager'],
    };
}

export function ProvisionMemberRequestToJSON(json: any): ProvisionMemberRequest {
    return ProvisionMemberRequestToJSONTyped(json, false);
}

export function ProvisionMemberRequestToJSONTyped(value?: ProvisionMemberRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'displayName': value['displayName'],
        'email': value['email'],
        'isAccountAdministrator': value['isAccountAdministrator'],
        'isEmployee': value['isEmployee'],
        'isManager': value['isManager'],
    };
}